# Mimir query engine: improve performance of `TestBothEnginesReturnSameResultsForBenchmarkQueries` and `TestBenchmarkSetup`