--- a/pkg/streamingpromql/benchmarks/ingester.go
+++ b/pkg/streamingpromql/benchmarks/ingester.go
@@ -224,16 +224,22 @@ func pushTestData(ing *ingester.Ingester, metricSizes []int) error {
 			end = NumIntervals
 		}
 
+		sampleCount := end - start
+
 		req := &mimirpb.WriteRequest{
 			Timeseries: make([]mimirpb.PreallocTimeseries, len(metrics)),
 		}
 
 		for metricIdx, m := range metrics {
+			series := mimirpb.PreallocTimeseries{TimeSeries: mimirpb.TimeseriesFromPool()}
+			series.Labels = mimirpb.FromLabelsToLabelAdapters(m.Copy())
+
 			if strings.HasPrefix(m.Get("__name__"), "nh_") {
-				series := mimirpb.PreallocTimeseries{TimeSeries: &mimirpb.TimeSeries{
-					Labels:     mimirpb.FromLabelsToLabelAdapters(m.Copy()),
-					Histograms: make([]mimirpb.Histogram, end-start),
-				}}
+				if cap(series.Histograms) < sampleCount {
+					series.Histograms = make([]mimirpb.Histogram, sampleCount)
+				}
+
+				series.Histograms = series.Histograms[:sampleCount]
 
 				for ts := start; ts < end; ts++ {
 					// TODO(jhesketh): Fix this with some better data
@@ -248,21 +254,20 @@ func pushTestData(ing *ingester.Ingester, metricSizes []int) error {
 					series.Histograms[ts-start].PositiveSpans = histogramSpans
 					series.Histograms[ts-start].PositiveDeltas = histogramDeltas
 				}
-
-				req.Timeseries[metricIdx] = series
 			} else {
-				series := mimirpb.PreallocTimeseries{TimeSeries: &mimirpb.TimeSeries{
-					Labels:  mimirpb.FromLabelsToLabelAdapters(m.Copy()),
-					Samples: make([]mimirpb.Sample, end-start),
-				}}
+				if cap(series.Samples) < sampleCount {
+					series.Samples = make([]mimirpb.Sample, sampleCount)
+				}
+
+				series.Samples = series.Samples[:sampleCount]
 
 				for ts := start; ts < end; ts++ {
 					series.Samples[ts-start].TimestampMs = int64(ts) * interval.Milliseconds()
 					series.Samples[ts-start].Value = float64(ts) + float64(metricIdx)/float64(len(metrics))
 				}
-
-				req.Timeseries[metricIdx] = series
 			}
+
+			req.Timeseries[metricIdx] = series
 		}
 		if _, err := ing.Push(ctx, req); err != nil {
 			return fmt.Errorf("failed to push samples to ingester: %w", err)