#### What this PR does

This PR improves the performance of two MQE tests, `TestBothEnginesReturnSameResultsForBenchmarkQueries` and `TestBenchmarkSetup`.

It does this by only flushing the ingester used during these tests once, rather than after loading each batch of data.

On my machine, this cuts ~30s off the time taken to run each of these tests:
* `TestBothEnginesReturnSameResultsForBenchmarkQueries`: was 50.1, now 20.2s
* `TestBenchmarkSetup`: was 36.1s, now 6.0s

This does not have a dramatic effect on the performance of setting up the test data used during benchmarks, as this is dominated by the time taken to ingest the data.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [n/a] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.