@charleskorn 2024-08-14
> Can you do a profile of the benchmark tool before and after this change please? I think it'll lead to OOMing in some cases.

Before this change (tl;dr: tests peak at 1.30 GB RSS, benchmark ingester peaks at 822 MB RSS):

```
pkg/streamingpromql/benchmarks$ go test -c -o tests -tags stringlabels . && time -l -h  ./tests -test.v -test.run="TestBothEnginesReturnSameResultsForBenchmarkQueries" -test.count=1

<snip>

  54.05s real   23.89s user   7.31s sys
          1303986176  maximum resident set size
                   0  average shared memory size
                   0  average unshared data size
                   0  average unshared stack size
               80118  page reclaims
                 485  page faults
                   0  swaps
                   0  block input operations
                   0  block output operations
                5052  messages sent
                5010  messages received
                7644  signals received
               62611  voluntary context switches
              180703  involuntary context switches
        285496870362  instructions retired
         77642138811  cycles elapsed
          1258557760  peak memory footprint


tools/benchmark-query-engine$ go build -tags stringlabels . && command time -l -h ./benchmark-query-engine -start-ingester

<snip>

	1m36.00s real		1m16.28s user		11.88s sys
           822231040  maximum resident set size
                   0  average shared memory size
                   0  average unshared data size
                   0  average unshared stack size
               54873  page reclaims
                 455  page faults
                   0  swaps
                   0  block input operations
                   0  block output operations
                   9  messages sent
                   9  messages received
               20670  signals received
               71493  voluntary context switches
             1647464  involuntary context switches
       1051284592543  instructions retired
        259614041582  cycles elapsed
           756141952  peak memory footprint

```

After this change (tl;dr: tests peak at 1.25 GB RSS, benchmark ingester peaks at 2.53 GB RSS):

```
pkg/streamingpromql/benchmarks$ go test -c -o tests -tags stringlabels . && time -l -h  ./tests -test.v -test.run="TestBothEnginesReturnSameResultsForBenchmarkQueries" -test.count=1

<snip>

	19.09s real		17.61s user		1.45s sys
          1252556800  maximum resident set size
                   0  average shared memory size
                   0  average unshared data size
                   0  average unshared stack size
               76403  page reclaims
                 485  page faults
                   0  swaps
                   0  block input operations
                   0  block output operations
                4851  messages sent
                4779  messages received
                3138  signals received
                9422  voluntary context switches
               50562  involuntary context switches
        244640504380  instructions retired
         57986569660  cycles elapsed
          1214075072  peak memory footprint

tools/benchmark-query-engine$ go build -tags stringlabels . && command time -l -h ./benchmark-query-engine -start-ingester

<snip>

       1m4.00s real		1m10.25s user		5.16s sys
          2534768640  maximum resident set size
                   0  average shared memory size
                   0  average unshared data size
                   0  average unshared stack size
              173994  page reclaims
                 455  page faults
                   0  swaps
                   0  block input operations
                   0  block output operations
                   9  messages sent
                   9  messages received
                6157  signals received
               15051  voluntary context switches
              214111  involuntary context switches
        816462636390  instructions retired
        230477153096  cycles elapsed
          2502367744  peak memory footprint

```

The main problem seems to be the benchmark setup case - I'll see if I can do anything to improve that.