@charleskorn 2024-08-14
I've tweaked the setup to use batches of 1000 samples, with a flush after every batch, and to reuse the `mimirpb.TimeSeries` instances and slices between batches.

This gives similar performance to flushing after all batches for the test cases, and seems to have made the benchmark setup time noticeably faster (previously it was over a minute, now it's consistently less than 45s). It also returns the peak RSS of the benchmark setup to ~800 MB.

@jhesketh could you give this branch a run on your machine and confirm you don't run into any OOMs?