@jhesketh 2024-08-14
> @jhesketh could you give this branch a run on your machine and confirm you don't run into any OOMs?

I ran it without OOM'ing. I didn't measure the memory usage, but watching top the memory seems to go up before flushing about ~1GB peaks.